% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-dkan_list.R
\name{as.dkan_list}
\alias{as.dkan_list}
\alias{is.dkan_list}
\title{dkan_list class helpers}
\usage{
as.dkan_list(x, ...)

is.dkan_list(x)
}
\arguments{
\item{x}{Variety of things, character, list, or dkan_list class object}

\item{...}{Further args passed on to \code{\link{retrieve_nodes}} if character given}
}
\description{
dkan_list class helpers
}
