% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-dkan_node.R
\name{as.dkan_node}
\alias{as.dkan_node}
\alias{is.dkan_node}
\title{dkan_node class helpers}
\usage{
as.dkan_node(x, ...)

is.dkan_node(x)
}
\arguments{
\item{x}{Variety of things, character, list, or dkan_node class object}

\item{...}{Further args passed on to \code{\link{retrieve_node}} if character given
# create item class from only an item ID
as.dkan_node("22")

# gives back itself
(x <- as.dkan_node("22"))
as.dkan_node(x)}
}
\description{
dkan_node class helpers
}
