% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_node.R
\name{create_node}
\alias{create_node}
\title{create_node}
\usage{
create_node(url = get_url(), body, credentials = list(cookie =
  dkanr::get_cookie(), token = dkanr::get_token()), as = "json", ...)
}
\arguments{
\item{url}{character: The DKAN site URL}

\item{body}{json: A JSON object to be passed to the body of the request}

\item{credentials}{Optional list parameter. Default values are Cookie and Token generated by dkan_setup()}

\item{as}{character: Output format. Options are: 'json' or 'list'}

\item{...}{Other optional parameters passed to the underlying GET request}
}
\value{
dkan_node object
}
\description{
Create a DKAN node
}
\examples{
\dontrun{
create_node(url = "http://demo.getdkan.com", body = {"title": "TEST DATASET"})
}
}
