% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_node.R
\name{delete_node}
\alias{delete_node}
\title{delete_node}
\usage{
delete_node(nid, url = get_url(), as = "json", ...)
}
\arguments{
\item{nid}{character: DKAN node ID}

\item{url}{character: The DKAN site URL}

\item{as}{character: Output format. Options are: 'json' or 'list'}

\item{...}{Other optional parameters passed to the underlying GET request}
}
\value{
character or dkan_node object
}
\description{
Delete an exisiting DAKN node
}
\examples{
delete_node(nid = 1, url = "http://demo.getdkan.com")
}
