% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkanr_settings.R
\name{dkanr_settings}
\alias{dkanr_settings}
\alias{get_url}
\alias{get_cookie}
\alias{get_token}
\title{Get or set dkanr DKAN settings}
\usage{
dkanr_settings()

get_url()

get_cookie()

get_token()
}
\value{
\code{dkanr_settings} prints your base url, API cookie and token (if used)
\code{dkanr_setup} sets your production settings
}
\description{
Get or set dkanr DKAN settings
}
\examples{
dkanr_settings()
}
\seealso{
\code{\link{dkanr_setup}},
\code{\link{get_default_url}}, \code{\link{get_default_username}},
\code{\link{get_default_password}}
}
