% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_is_available.R
\name{ds_is_available}
\alias{ds_is_available}
\title{ds_is_available}
\usage{
ds_is_available(metadata)
}
\arguments{
\item{metadata}{list: Resource level metadata}
}
\value{
logical vector
}
\description{
Returns TRUE is a data file has been imported in the DKAN datastore, returns FALSE otherwise
}
