% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filters_to_text_query}
\alias{filters_to_text_query}
\title{filters_to_text_query}
\usage{
filters_to_text_query(filters, text)
}
\arguments{
\item{filters}{named vector: filters to be applied}

\item{text}{vector: The key word used for filtering}
}
\description{
Helper function to build API query string
}
