% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resource_nids.R
\name{get_resource_nids}
\alias{get_resource_nids}
\title{get_resource_nids}
\usage{
get_resource_nids(metadata)
}
\arguments{
\item{metadata}{list: Dataset level metadata}
}
\description{
Returns all the resource node ids attached to a specific node id
}
