% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resource_url.R
\name{get_resource_url}
\alias{get_resource_url}
\title{get_resource_url}
\usage{
get_resource_url(metadata)
}
\arguments{
\item{metadata}{list: Resource level metadata}
}
\value{
character vector
}
\description{
Returns the download link or external page URL for a specific resource
}
