% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-dkan_list.R
\name{is.dkan_list}
\alias{is.dkan_list}
\title{dkan_list class helpers}
\usage{
is.dkan_list(x)
}
\arguments{
\item{x}{Variety of things, character, list, or dkan_list class object
as.dkan_list <- function(x, ...) UseMethod("as.dkan_list")

#' @export
as.dkan_list.character <- function(x, ...) retrieve_node(x, ...)

#' @export
as.dkan_list.dkan_list <- function(x, ...) x

#' @export
as.dkan_list.list <- function(x, ...) structure(x, class = "dkan_list")}
}
\description{
dkan_list class helpers
}
