% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-dkan_node.R
\name{is.dkan_node}
\alias{is.dkan_node}
\title{dkan_node class helpers}
\usage{
is.dkan_node(x)
}
\arguments{
\item{x}{Variety of things, character, list, or dkan_node class object
# create item class from only an item ID
as.dkan_node("22")

# gives back itself
(x <- as.dkan_node("22"))
as.dkan_node(x)
as.dkan_node <- function(x, ...) UseMethod("as.dkan_node")

#' @export
as.dkan_node.character <- function(x, ...) retrieve_node(x, ...)

#' @export
as.dkan_node.dkan_node <- function(x, ...) x

#' @export
as.dkan_node.list <- function(x, ...) structure(x, class = "dkan_node")}
}
\description{
dkan_node class helpers
}
