% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_all_nodes.R
\name{list_all_nodes}
\alias{list_all_nodes}
\title{list_all_nodes}
\usage{
list_all_nodes(url = get_url(), fields = NULL, filters = NULL,
  as = "json", ...)
}
\arguments{
\item{url}{character: The DKAN site URL}

\item{fields}{character vector: fields to be returned by the request}

\item{filters}{named character vector: filters to be applied to the search}

\item{as}{character: Output format. Options are: 'json' or 'list'}

\item{...}{Other optional parameters to be passed to the underlying GET request}
}
\value{
dkan_node object
}
\description{
Returns list of nodes
}
\examples{
list_all_nodes(url = "http://demo.getdkan.com")
list_all_nodes(url = "http://demo.getdkan.com", fields = c('nid', 'type', 'uri'), filters = c(type='resource'))
}
