% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_nodes_all.R
\name{list_nodes_all}
\alias{list_nodes_all}
\title{list_nodes_all}
\usage{
list_nodes_all(url = get_url(), credentials = list(cookie =
  dkanr::get_cookie(), token = dkanr::get_token()), fields = NULL,
  filters = NULL, as = "df", ...)
}
\arguments{
\item{url}{character: The DKAN site URL}

\item{credentials}{Optional list parameter. Default values are Cookie and Token generated by dkan_setup()}

\item{fields}{character vector: fields to be returned by the request}

\item{filters}{named character vector: filters to be applied to the search}

\item{as}{character: Output format. Options are: 'json', 'list' or 'df' (data frame)}

\item{...}{Other optional parameters to be passed to the underlying GET request}
}
\value{
json, dkan_node (list), or data frame
}
\description{
Returns list of nodes
}
\examples{
\dontrun{
list_nodes_all(url = "http://demo.getdkan.com")
list_nodes_all(url = "http://demo.getdkan.com",
fields = c('nid', 'type'),
filters = c(type = 'resource'))
}
}
