% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-login_service_json.R
\docType{data}
\name{login_service_json}
\alias{login_service_json}
\title{JSON body for login service request}
\format{A list of 3.}
\usage{
login_service_json
}
\description{
JSON body for login service request
}
\examples{
login_service_json
}
\keyword{datasets}
