% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_node.R
\name{retrieve_node}
\alias{retrieve_node}
\title{retrieve_node}
\usage{
retrieve_node(nid, url = get_url(), credentials = list(cookie =
  dkanr::get_cookie(), token = dkanr::get_token()), as = "json", ...)
}
\arguments{
\item{nid}{character: DKAN node ID}

\item{url}{character: The DKAN site URL}

\item{credentials}{Optional list parameter. Default values are Cookie and Token generated by dkan_setup()}

\item{as}{character: Output format. Options are: 'json' or 'list'}

\item{...}{Other optional parameters passed to the underlying GET request}
}
\value{
dkan_node object
}
\description{
Returns metadata associated with an exisiting DKAN Node ID.
}
\examples{
retrieve_node(nid = 1, url = "http://demo.getdkan.com")
}
