% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-system_connect_json.R
\docType{data}
\name{system_connect_json}
\alias{system_connect_json}
\title{JSON body for system connect request}
\format{A list of 3.}
\usage{
system_connect_json
}
\description{
JSON body for system connect request
}
\examples{
system_connect_json
}
\keyword{datasets}
