% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_node.R
\name{update_node}
\alias{update_node}
\title{update_node}
\usage{
update_node(nid, url = get_url(), body, credentials = list(cookie =
  dkanr::get_cookie(), token = dkanr::get_token()), as = "json", ...)
}
\arguments{
\item{nid}{character: DKAN node ID}

\item{url}{character: The DKAN site URL}

\item{body}{json: A JSON object to be passed to the body of the request}

\item{credentials}{Optional list parameter. Default values are Cookie and Token generated by dkan_setup()}

\item{as}{character: Output format. Options are: 'json' or 'list'}

\item{...}{Other optional parameters passed to the underlying GET request}
}
\value{
dkan_node object
}
\description{
Update an exisiting DAKN node
}
\examples{
\dontrun{
update_node(nid = 1, url = "http://demo.getdkan.com")
}
}
