% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkanr_logout.R
\name{dkanr_logout}
\alias{dkanr_logout}
\title{dkanr_logout}
\usage{
dkanr_logout(url = get_url(), credentials = list(cookie =
  dkanr::get_cookie(), token = dkanr::get_token()))
}
\arguments{
\item{url}{Character: API root URL. Default value is the url returned by get_url()}

\item{credentials}{Optional list parameter. Default values are Cookie and Token generated by dkan_setup()}
}
\value{
json
}
\description{
Close connection with DKAN API initiated by dkanr_setup()
}
