% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_search.R
\name{ds_search}
\alias{ds_search}
\title{ds_search}
\usage{
ds_search(resource_id = "10c578a6-63c4-40bd-a55d-0c27bf276283",
  fields = NULL, filters = NULL, num_records = NULL, offset = 0,
  sort_by = NULL, q = NULL, url = get_url(), credentials = list(cookie =
  dkanr::get_cookie(), token = dkanr::get_token()), as = "df")
}
\arguments{
\item{resource_id}{character: DKAN resource ID}

\item{fields}{character vector: fields to be returned by the request}

\item{filters}{named character vector: filters to be applied when retrieving records}

\item{num_records}{numeric: Number of records to retrieve}

\item{offset}{numeric: Number of results to skip in the beginning}

\item{sort_by}{named character vector: field to be sorted by and the order of sorting}

\item{q}{character: full text query filter}

\item{url}{character: The DKAN site URL}

\item{as}{character: output format - json or df}

\item{limit}{numeric: Number of results to retrieve per page}
}
\value{
data.frame or character
}
\description{
Search csv resource file stored in the datastore.
}
\examples{
ds_search(resource_id = '10c578a6-63c4-40bd-a55d-0c27bf276283',
}
